+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Implementation of the ambiguity confidence measure. [#162] 
+
 ### BREAKING CHANGE
 
 - GeoRef is now supported and both `crs` & `transform` ought to be set as attributes of images dataset provided to Pandora [#158
@@ -39,6 +43,7 @@
 ### Fixed
 
 - Confidence measure that computes the distance LR / RL in cross checking method. [#155]
+
 
 ## 0.4.0 (November 2020)